// domains-frontend/src/api/models/domain-order/domainResponse.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface DomainResponse {
  id?: string;
  fqdn?: string;
  zoneId?: string;
  zoneName?: string;
  activatedAt?: string;
  expiresAt?: string;
  createdAt?: string;
  updatedAt?: string;
}